+++ tests/e2e-leg-9/replicated-upgrade-sanity/setup-vdb/base/setup-vdb.yaml
@@ -17,11 +17,12 @@ metadata:
   name: v-base-upgrade
   annotations:
     vertica.com/include-uid-in-path: true
+    vertica.com/is-new-online-upgrade: true
 spec:
   image: kustomize-vertica-image
   dbName: repUP
   initPolicy: CreateSkipPackageInstall
-  upgradePolicy: Replicated
+  upgradePolicy: Online
   communal: {}
   local:
     requestSize: 250Mi
